Raspi-Kiosk/Portland/requirements.txt
Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1 